Create Customer Retention Analysis For Airtel Telecoms (Demo)
This is a Project on Customer Retention that captures adequate KPI's to drive insight into the major areas to look into as to why the churn rate of clients were quite high the last month. Very clear visualization was done and displayed on the dashboard using Power BI from which intelligent business decision will be taken.
Some of the things to expect includes ; Total Churn, Churn rate, Customer Demography, Customer Satisfaction, Service Quality and Resolved Tickets.